Containers/Crawler/requirements.txt
yfinance
mysql-connector-python
grpcio
protobuf
kafka-python
six